feat(io): add --writer-key-file flag to load writer encryption keys from JSON
